refactor: can we simplify this even further??
